requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Dropbox Integration
dropbox==11.36.2

# AMQP Messaging
pika==1.3.2
celery==5.3.4
redis==5.0.1

# Security & Encryption
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP & Webhooks
httpx==0.25.2
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
